requirements.txt
# ZEEPT - LLM Fine-tuning and Evaluation Platform
# Requirements file for pip installation

# Core ML/DL
torch>=2.0.0
numpy>=1.24.0

# Tokenization
tiktoken>=0.5.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data Processing
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: Progress bars
tqdm>=4.65.0
